Refactor app architecture with navigation and domain layers
