+++ app/src/main/java/com/gerardo/comandas/data/repository/ComandaRepository.kt
@@ -1,0 +1,7 @@
+package com.gerardo.comandas.data.repository
+
+import com.gerardo.comandas.domain.model.Producto
+
+interface ComandaRepository {
+    fun getProductos(): List<Producto>
+}
