+++ app/src/main/java/com/gerardo/comandas/printing/Printer.kt
@@ -1,0 +1,3 @@
+package com.gerardo.comandas.printing
+
+class Printer
