+++ app/src/main/java/com/gerardo/comandas/sync/SyncManager.kt
@@ -1,0 +1,3 @@
+package com.gerardo.comandas.sync
+
+class SyncManager
