+++ app/src/main/java/com/gerardo/comandas/ui/components/ComidaYBebidaButtons.kt
@@ -1,0 +1,37 @@
+package com.gerardo.comandas.ui.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+
+@Composable
+fun ComidaYBebidaButtons(navController: NavController) {
+    Row(
+        horizontalArrangement = Arrangement.SpaceEvenly,
+        modifier = Modifier.fillMaxWidth().padding(16.dp)
+    ) {
+        Button(
+            onClick = { navController.navigate("comida_screen") },
+            colors = ButtonDefaults.buttonColors(containerColor = Color(0xFFFFC107)),
+            modifier = Modifier.padding(8.dp)
+        ) {
+            Text(text = "Comida", color = Color.Black)
+        }
+        Button(
+            onClick = { navController.navigate("bebida_screen") },
+            colors = ButtonDefaults.buttonColors(containerColor = Color(0xFF03A9F4)),
+            modifier = Modifier.padding(8.dp)
+        ) {
+            Text(text = "Bebida", color = Color.Black)
+        }
+    }
+}
