+++ app/src/main/java/com/gerardo/comandas/ui/screens/IngredientesScreen.kt
@@ -1,0 +1,37 @@
+package com.gerardo.comandas.ui.screens
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.gerardo.comandas.ui.components.PantallaConRibetes
+
+@Composable
+fun IngredientesScreen(navController: NavController, nombre: String, ingredientes: String) {
+    PantallaConRibetes(navController = navController) {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(Color(0xFFC7F3E3)),
+            contentAlignment = Alignment.Center
+        ) {
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Center
+            ) {
+                Text(text = nombre, style = MaterialTheme.typography.headlineMedium)
+                Text(
+                    text = ingredientes,
+                    style = MaterialTheme.typography.bodyLarge,
+                    modifier = Modifier.padding(16.dp)
+                )
+            }
+        }
+    }
+}
