+++ app/src/main/java/com/gerardo/comandas/ui/screens/MainScreen.kt
@@ -1,0 +1,55 @@
+package com.gerardo.comandas.ui.screens
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.gerardo.comandas.R
+
+@Composable
+fun MainScreen(modifier: Modifier = Modifier, navController: NavController) {
+    Box(
+        modifier = modifier
+            .fillMaxSize()
+            .background(Color(0xFFE573A6)),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center
+        ) {
+            Image(
+                painter = painterResource(id = R.drawable.logo_mylittlediner),
+                contentDescription = "Logo My Little Diner",
+                modifier = Modifier
+                    .size(220.dp)
+                    .padding(16.dp)
+                    .clickable {
+                        navController.navigate("login_screen")
+                    },
+                contentScale = ContentScale.Fit
+            )
+            Spacer(modifier = Modifier.height(24.dp))
+            Text(
+                text = "MY LITTLE DINER",
+                style = MaterialTheme.typography.headlineMedium.copy(
+                    fontWeight = FontWeight.Bold,
+                    color = Color.White
+                ),
+                textAlign = TextAlign.Center
+            )
+        }
+    }
+}
