+++ app/src/main/java/com/gerardo/comandas/utils/Constants.kt
@@ -1,0 +1,3 @@
+package com.gerardo.comandas.utils
+
+object Constants
